# ops-ui/openapi/codex-api.openapi.json
{
  "components": {
    "schemas": {
...
      "DeleteResponse": {
        "properties": {
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "DeleteResponse",
        "type": "object"
      },
      "DomainEntryAliasCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "aliasKind": {
            "default": "ALIAS",
            "enum": [
              "ALIAS",
              "ASR_ERROR",
              "SEARCH_ALIAS",
              "NICKNAME",
              "WORDPLAY",
              "MISSPELLING"
            ],
            "title": "Aliaskind",
            "type": "string"
          },
          "applyScope": {
            "default": "SEARCH_ONLY",
            "enum": [
              "NONE",
              "SEARCH_ONLY",
              "SEARCH_AND_SUMMARY",
              "DISPLAY_ALLOWED"
            ],
            "title": "Applyscope",
            "type": "string"
          },
          "certainty": {
            "default": "MEDIUM",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "title": "Certainty",
            "type": "string"
          },
          "languageCode": {
            "anyOf": [
              {
                "maxLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Languagecode"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "surfaceForm": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Surfaceform",
            "type": "string"
          }
        },
        "required": [
          "surfaceForm"
        ],
        "title": "DomainEntryAliasCreateRequest",
        "type": "object"
      },
...
      "DomainEntryCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "aliases": {
            "items": {
              "$ref": "#/components/schemas/DomainEntryAliasCreateRequest"
            },
            "title": "Aliases",
            "type": "array"
          },
          "canonicalName": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Canonicalname",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "disambiguation": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disambiguation"
          },
          "displayName": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "isActive": {
            "default": true,
            "title": "Isactive",
            "type": "boolean"
          },
          "priority": {
            "default": 50,
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Priority",
            "type": "integer"
          },
          "promptPolicy": {
            "default": "AUTO_ON_MATCH",
            "enum": [
              "AUTO_ON_MATCH",
              "ALWAYS_FOR_SCOPED_STREAMER",
              "DISABLED"
            ],
            "title": "Promptpolicy",
            "type": "string"
          },
          "sourceNote": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcenote"
          },
          "streamerIds": {
            "items": {
              "type": "integer"
            },
            "title": "Streamerids",
            "type": "array"
          },
          "typeId": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typeid"
          },
          "typeKey": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typekey"
          },
          "typeLabel": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typelabel"
          }
        },
        "required": [
          "canonicalName"
        ],
        "title": "DomainEntryCreateRequest",
        "type": "object"
      },
      "DomainEntryListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DomainEntryResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "DomainEntryListResponse",
        "type": "object"
      },
      "DomainEntryResponse": {
        "properties": {
          "aliases": {
            "items": {
              "$ref": "#/components/schemas/DomainEntryAliasResponse"
            },
            "title": "Aliases",
            "type": "array"
          },
          "canonicalName": {
            "title": "Canonicalname",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "disambiguation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disambiguation"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "entryId": {
            "title": "Entryid",
            "type": "integer"
          },
          "isActive": {
            "title": "Isactive",
            "type": "boolean"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "promptPolicy": {
            "enum": [
              "AUTO_ON_MATCH",
              "ALWAYS_FOR_SCOPED_STREAMER",
              "DISABLED"
            ],
            "title": "Promptpolicy",
            "type": "string"
          },
          "sourceNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcenote"
          },
          "streamers": {
            "items": {
              "$ref": "#/components/schemas/DomainEntryStreamerResponse"
            },
            "title": "Streamers",
            "type": "array"
          },
          "typeId": {
            "title": "Typeid",
            "type": "integer"
          },
          "typeKey": {
            "title": "Typekey",
            "type": "string"
          },
          "typeLabel": {
            "title": "Typelabel",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "entryId",
          "typeId",
          "typeKey",
          "typeLabel",
          "canonicalName",
          "displayName",
          "disambiguation",
          "detail",
          "promptPolicy",
          "priority",
          "isActive",
          "sourceNote",
          "streamers",
          "aliases",
          "createdAt",
          "updatedAt"
        ],
        "title": "DomainEntryResponse",
        "type": "object"
      },
      "DomainEntryStreamerLinkRequest": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "relevance": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevance"
          },
          "streamerId": {
            "minimum": 1.0,
            "title": "Streamerid",
            "type": "integer"
          }
        },
        "required": [
          "streamerId"
        ],
        "title": "DomainEntryStreamerLinkRequest",
        "type": "object"
      },
...
      "DomainEntryUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "canonicalName": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonicalname"
          },
          "detail": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "disambiguation": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disambiguation"
          },
          "displayName": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "isActive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isactive"
          },
          "priority": {
            "anyOf": [
              {
                "maximum": 1000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "promptPolicy": {
            "anyOf": [
              {
                "enum": [
                  "AUTO_ON_MATCH",
                  "ALWAYS_FOR_SCOPED_STREAMER",
                  "DISABLED"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promptpolicy"
          },
          "sourceNote": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcenote"
          },
          "typeId": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typeid"
          },
          "typeKey": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typekey"
          },
          "typeLabel": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typelabel"
          }
        },
        "title": "DomainEntryUpdateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "REST API wrapper for Codex SDK run/account/logout workflows.",
    "title": "Codex SDK CLI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/domain-entries": {
      "get": {
        "operationId": "list_domain_entries_domain_entries_get",
        "parameters": [
          {
            "in": "query",
            "name": "streamerId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Streamerid"
            }
          },
          {
            "in": "query",
            "name": "typeId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Typeid"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": true,
              "title": "Active"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainEntryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Domain Entries",
        "tags": [
          "domain-knowledge"
        ]
      },
      "post": {
        "operationId": "create_domain_entry_domain_entries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainEntryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Domain Entry",
        "tags": [
          "domain-knowledge"
        ]
      }
    },
    "/domain-entries/{entry_id}": {
      "delete": {
        "operationId": "archive_domain_entry_domain_entries__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Domain Entry",
        "tags": [
          "domain-knowledge"
        ]
      },
      "get": {
        "operationId": "get_domain_entry_domain_entries__entry_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Domain Entry",
        "tags": [
          "domain-knowledge"
        ]
      },
      "patch": {
        "operationId": "update_domain_entry_domain_entries__entry_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainEntryUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Domain Entry",
        "tags": [
          "domain-knowledge"
        ]
      }
    },
    "/domain-entries/{entry_id}/aliases": {
      "post": {
        "operationId": "add_domain_entry_alias_domain_entries__entry_id__aliases_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainEntryAliasCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Domain Entry Alias",
        "tags": [
          "domain-knowledge"
        ]
      }
    },
    "/domain-entries/{entry_id}/streamers": {
      "post": {
        "operationId": "add_domain_entry_streamer_domain_entries__entry_id__streamers_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainEntryStreamerLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Domain Entry Streamer",
        "tags": [
          "domain-knowledge"
        ]
      }
    },
    "/domain-entries/{entry_id}/streamers/{streamer_id}": {
      "delete": {
        "operationId": "remove_domain_entry_streamer_domain_entries__entry_id__streamers__streamer_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Entry Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "streamer_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Streamer Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Domain Entry Streamer",
        "tags": [
          "domain-knowledge"
        ]
      }
    },
    "/domain-entry-aliases/{alias_id}": {
      "delete": {
        "operationId": "delete_domain_entry_alias_domain_entry_aliases__alias_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "alias_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Alias Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Domain Entry Alias",
        "tags": [
          "domain-knowledge"
        ]
      },
      "patch": {
        "operationId": "update_domain_entry_alias_domain_entry_aliases__alias_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "alias_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Alias Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainEntryAliasUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainEntryAliasResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Domain Entry Alias",
        "tags": [
          "domain-knowledge"
        ]
      }
    },
    "/domain-entry-types": {
      "get": {
        "operationId": "list_domain_entry_types_domain_entry_types_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DomainEntryTypeResponse"
                  },
                  "title": "Response List Domain Entry Types Domain Entry Types Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Domain Entry Types",
        "tags": [
          "domain-knowledge"
        ]
      },
      "post": {
        "operationId": "create_domain_entry_type_domain_entry_types_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainEntryTypeCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainEntryTypeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Domain Entry Type",
        "tags": [
          "domain-knowledge"
        ]
      }
    },
...
  }
}